implement review suggestions + fix builds
