+++ octavia-cli/integration_tests/test_generate/conftest.py
@@ -11,8 +11,3 @@ def octavia_project_directory(tmpdir):
     for directory in OCTAVIA_PROJECT_DIRECTORIES:
         tmpdir.mkdir(directory)
     return tmpdir
-
-
-@pytest.fixture
-def mock_api_client(mocker):
-    return mocker.Mock()
