+++ octavia-cli/integration_tests/test_generate/expected_rendered_yaml/destination_s3/input_spec.yaml
