+++ octavia-cli/setup.py
@@ -40,7 +40,7 @@ setup(
         "Source": "https://github.com/airbytehq/airbyte",
         "Tracker": "https://github.com/airbytehq/airbyte/issues",
     },
-    packages=find_packages(exclude=("unit_tests", "docs")),
+    packages=find_packages(exclude=("unit_tests", "integration_tests", "docs")),
     install_requires=[
         "click~=8.0.3",
         f"airbyte_api_client @ file://{os.getcwd()}/build/airbyte_api_client",
@@ -50,6 +50,7 @@ setup(
     python_requires=">=3.8.12",
     extras_require={
         "dev": ["MyPy~=0.812", "pytest~=6.2.5", "pytest-cov", "pytest-mock", "requests-mock", "pre-commit"],
+        "tests": ["MyPy~=0.812", "pytest~=6.2.5", "pytest-cov", "pytest-mock", "requests-mock", "pre-commit"],
         "sphinx-docs": [
             "Sphinx~=4.2",
             "sphinx-rtd-theme~=1.0",
