+++ octavia-cli/unit_tests/conftest.py
@@ -1,0 +1,10 @@
+#
+# Copyright (c) 2021 Airbyte, Inc., all rights reserved.
+#
+
+import pytest
+
+
+@pytest.fixture
+def mock_api_client(mocker):
+    return mocker.Mock()
